Collect dt_control_log() and dt_toast_log() messages
Visibility of messges shown via dt_control_log(), dt_toast_log() and dt_toast_markup_log() is currently
pretty bad as only the last shown message is shown if new ones are "pushed".

This commit redesigns both buffers to be used as circular buffers, a new call to any function above
will not display the last text but all available and not-yet ackknowledged (or disappered) messages
for improved user readability.